src/RPi/neopixels/requirements.txt
paho-mqtt
pyyaml
neopixel
